# API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mangette v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/Chapters/Manga/{MangaId}": {
      "post": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns all API.Schema.MangaContext.Chapter of API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page to request (default 1)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Size of Page (default 10)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterPagedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Page data wrong",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database request"
          }
        }
      }
    },
    "/v2/Chapters/LatestAvailable/{MangaId}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns the latest API.Controllers.DTOs.Chapter of requested API.Schema.MangaContext.Manga",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No available chapters"
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found.",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/Chapters/LatestDownloaded/{MangaId}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns the latest API.Controllers.DTOs.Chapter of requested API.Schema.MangaContext.Manga that is downloaded",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "204": {
            "description": "No available chapters"
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found.",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "412": {
            "description": "Could not retrieve the maximum chapter-number",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Retry after timeout, updating value"
          }
        }
      }
    },
    "/v2/Chapters/IgnoreBefore/{MangaId}": {
      "patch": {
        "tags": [
          "Chapters"
        ],
        "summary": "Configure the API.Controllers.DTOs.Chapter cut-off for API.Schema.MangaContext.Manga",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Threshold (API.Controllers.DTOs.Chapter ChapterNumber)",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "number",
                "format": "float"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "number",
                "format": "float"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "number",
                "format": "float"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "number",
                "format": "float"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found.",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "202": {
            "description": ""
          }
        }
      }
    },
    "/v2/Chapters/{ChapterId}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns API.Controllers.DTOs.Chapter with ChapterId",
        "parameters": [
          {
            "name": "ChapterId",
            "in": "path",
            "description": "API.Controllers.DTOs.Chapter.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.Chapter with ChapterId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chapters"
        ],
        "summary": "Deletes API.Controllers.DTOs.Chapter with ChapterId",
        "parameters": [
          {
            "name": "ChapterId",
            "in": "path",
            "description": "API.Controllers.DTOs.Chapter.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "API.Controllers.DTOs.Chapter with ChapterId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/Chapters/ConnectorId/{MangaConnectorIdId}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns the API.Controllers.DTOs.MangaConnectorId`1 with API.Controllers.DTOs.MangaConnectorId`1.Key",
        "parameters": [
          {
            "name": "MangaConnectorIdId",
            "in": "path",
            "description": "Key of API.Controllers.DTOs.MangaConnectorId`1",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterMangaConnectorId"
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.MangaConnectorId`1 with MangaConnectorIdId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chapters"
        ],
        "summary": "Deletes the API.Controllers.DTOs.MangaConnectorId`1 with API.Controllers.DTOs.MangaConnectorId`1.Key",
        "parameters": [
          {
            "name": "MangaConnectorIdId",
            "in": "path",
            "description": "Key of API.Controllers.DTOs.MangaConnectorId`1",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "API.Controllers.DTOs.MangaConnectorId`1 with MangaConnectorIdId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/Chapters/{ChapterId}/DownloadFrom/{MangaConnectorName}/{IsRequested}": {
      "patch": {
        "tags": [
          "Chapters"
        ],
        "summary": "(Un-)Marks API.Controllers.DTOs.Chapter as requested for Download from API.MangaConnectors.MangaConnector",
        "parameters": [
          {
            "name": "ChapterId",
            "in": "path",
            "description": "API.Controllers.DTOs.Chapter with ChapterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MangaConnectorName",
            "in": "path",
            "description": "API.MangaConnectors.MangaConnector with MangaConnectorName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsRequested",
            "in": "path",
            "description": "true to mark as requested, false to mark as not-requested",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "ChapterId or MangaConnectorName not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "428": {
            "description": "API.Controllers.DTOs.Chapter is not linked to API.MangaConnectors.MangaConnector yet. Search for API.Controllers.DTOs.Chapter on API.MangaConnectors.MangaConnector first (to create a API.Controllers.DTOs.MangaConnectorId`1).",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/Manga/{MangaId}": {
      "get": {
        "tags": [
          "Manga"
        ],
        "summary": "Return API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Manga"
        ],
        "summary": "Delete API.Controllers.DTOs.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Controllers.DTOs.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "API.Controllers.DTOs.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/Manga/ForceRecheck/Chapter/{chapterId}": {
      "post": {
        "tags": [
          "Manga"
        ],
        "summary": "Force re-check a specific API.Schema.MangaContext.Chapter by deleting its record.",
        "parameters": [
          {
            "name": "chapterId",
            "in": "path",
            "description": "API.Schema.MangaContext.Chapter.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Affected records",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChapterFilterRecord": {
        "type": "object",
        "properties": {
          "downloaded": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "volumeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "chapterNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChapterPagedResponse": {
        "required": [
          "data",
          "page",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chapter"
            },
            "description": "The data"
          },
          "page": {
            "type": "integer",
            "description": "The page-number",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages (max page)",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "The total number of entries for this request",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}